2.0 DIP fix.

+++ qups-guard.cpp
@@ -11,10 +11,15 @@ struct gpiod_line *linePfo;
 struct gpiod_line *lineLim;
 struct gpiod_line *lineShd;
 u_int8_t lastval_pfo = 255, lastval_lim = 255;
+u_int8_t shutdown_delay = 0;
+char dip_sw[3];
 struct timespec ts;
 
 #define CONSUMER "qUPS-guard"
 #define POLLINTERVAL 500000
+#define DEBOUNCE_COUNT 100
+#define DEBOUNCE_INTERVAL 10000
+#define SHUTDOWN_DELAY 10
 
 struct DIPsw
 {
@@ -40,6 +45,24 @@ struct DIPsw DIPswa[10] = {
 {"111", 4, 24, 23}, {"011", 14, 18, 15}, {"101", 25, 7, 8}, {"001", 17, 22, 27}, {"110", 10, 11, 9}, {"010", 12, 20, 16}, {"100", 19, 21, 26}
 };
 
+u_int8_t debounce_limit() {
+    u_int8_t debcount=0;
+    while (++debcount < DEBOUNCE_COUNT) {
+        if (gpiod_line_get_value(lineLim)) {
+            // lineLim bouncing
+            syslog(LOG_INFO, "Limit low returned on %d/%d\n", debcount, DEBOUNCE_COUNT);
+            return 1;
+        }
+        usleep(DEBOUNCE_INTERVAL);
+    }
+    syslog(LOG_INFO, "Limit low not returned - initiate shutdown.\n", debcount);
+    u_int8_t shdcnt=shutdown_delay;
+    while (shdcnt-->0) {
+        syslog(LOG_INFO, "Shutdown delay counter: %d/%d.\n", shdcnt, shutdown_delay);
+	sleep(1);
+    }
+    return 0;
+}
 
 void SM()
 {
@@ -53,10 +76,10 @@ void SM()
             {
             case 0:
                 syslog(LOG_INFO, "Power NOK!");
-                continue;
+                break;
             case 1:
                 syslog(LOG_INFO, "Power OK.");
-                continue;
+                break;
             }
         }
 
@@ -67,15 +90,19 @@ void SM()
             switch (lim_val)
             {
             case 0:
-                syslog(LOG_INFO, "UPS level LOW - initiating shutdown seqence.");
+                syslog(LOG_INFO, "UPS level LOW - prepare shutdown seqence.");
+		if (debounce_limit()) {
+                    syslog(LOG_INFO, "Limit debounce - shutdown postponed.");
+		    continue;
+		}
                 if (system("sudo shutdown -h now") == 0)
                 {
                     syslog(LOG_INFO, "Shutdown sequence succesfully initiated.");
                 }
-                continue;
+                break;
             case 1:
                 syslog(LOG_INFO, "UPS level HIGH.");
-                continue;
+                break;
             }
         }
         usleep(POLLINTERVAL);
@@ -119,46 +146,65 @@ int g_gpioinit()
     {
         syslog(LOG_ERR, "Lim line event request failed: %s", strerror(errno));
     }
-
     return 0;
 }
 
 
 int main(int argc, char **argv)
 {
+    bool mat=false, mbt=false;
     openlog(CONSUMER, LOG_PID | LOG_NDELAY, LOG_USER);
-    if (argc == 2)
-    {
-        syslog(LOG_INFO, "Input argument: %s", argv[1]);
-        bool mat = false;
-        for (u_int8_t i = 0; i < 10; i++)
-        {
-            if (!strcmp(argv[1], DIPswa[i].DIP))
-            {
-		DIP_sw=DIPswa[i];
-                mat = true;
+
+    for (u_int8_t i = 1; i < argc; i++) {
+        if (strcmp(argv[i], "--shutdown-delay") == 0) {
+            mbt=true;
+            if (i + 1 < argc) {
+                shutdown_delay = atoi(argv[i + 1]);
+                i++;
+            } else {
+                fprintf(stderr, "Error: --shutdown-delay requires an argument - using default %d.\n", SHUTDOWN_DELAY);
+                shutdown_delay = SHUTDOWN_DELAY;
             }
-        }
-        if (mat)
-        {
-            syslog(LOG_INFO, "Used pins (BCM) - pfo: %d, lim: %d, shd: %d", DIP_sw.pfo_n, DIP_sw.lim_n, DIP_sw.shd_n);
-        }
-        else
-        {
-            syslog(LOG_ERR, "No match found!");
-            exit(-1);
+        } else if (strcmp(argv[i], "--dip") == 0) {
+            if (i + 1 < argc) {
+		u_int8_t dip_len;
+		dip_len = strlen(argv[i + 1]);
+		if (dip_len == 3 || dip_len == 2) {
+	                strncpy(dip_sw, argv[i + 1], dip_len);
+			for (u_int8_t j=0; j<10; j++) {
+				if (!strcmp(dip_sw, DIPswa[j].DIP)) {
+					DIP_sw=DIPswa[j];
+					mat=true;
+				}
+			}
+		}
+                i++;
+            } else {
+                fprintf(stderr, "Error: --dip requires an argument\n");
+                exit(EXIT_FAILURE);
+            }
+        } else {
+            fprintf(stderr, "Unknown argument: %s\n", argv[i]);
+            exit(EXIT_FAILURE);
         }
     }
-    else if (argc < 2)
-    {
-        syslog(LOG_ERR, "Too few arguments...#1-3 (P1-P2-P3) or #1-2 (GT1-GT2) DIP binary pattern needed.");
-        exit(0);
+
+    if (mat) {
+	syslog(LOG_INFO, "Used pins (BCM) - pfo: %d, lim: %d, shd: %d", DIP_sw.pfo_n, DIP_sw.lim_n, DIP_sw.shd_n);
+	printf("Used pins (BCM) - pfo: %d, lim: %d, shd: %d\n", DIP_sw.pfo_n, DIP_sw.lim_n, DIP_sw.shd_n);
+    } else {
+	syslog(LOG_ERR, "Used pins not specified - usage: --dip <DIP switch GT1-2 or DIP 1-2-3>");
+	syslog(LOG_ERR, "Example: --dip 10  means DIP switch GT1=ON GT2=OFF");
+	syslog(LOG_ERR, "         --dip 100 means DIP 1=ON 2=OFF 3=OFF");
+
+	printf("Used pins not specified - usage: --dip <DIP switch GT1-2 or DIP 1-2-3>\n");
+	printf("Example: --dip 10  means DIP switch GT1=ON GT2=OFF\n");
+	printf("         --dip 100 means DIP 1=ON 2=OFF 3=OFF\n");
     }
-    else if (argc > 2)
-    {
-        syslog(LOG_ERR, "Too many arguments...#1-3 (P1-P2-P3) or #1-2 (GT1-GT2) DIP binary pattern needed.");
-        exit(0);
+    if (!mbt) {
+	shutdown_delay=SHUTDOWN_DELAY;
     }
+    syslog(LOG_INFO, "Shutdown delay %d", shutdown_delay);
 
     if (!g_gpioinit())
     {
